+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from './types';
-export { Store } from './store';
-export { Model } from './model';
-export { Query } from './query';
+export * from './store';
+export * from './model';
+export * from './query';