+++ src/model.ts
@@ -1,9 +1,5 @@
 import { Store } from './store';
-import {
-    JsonApiIdentifier,
-    JsonApiRelationships,
-    JsonApiResource, KeyValueObject
-} from './types';
+import { JsonApiIdentifier, JsonApiRelationships, JsonApiResource, KeyValueObject } from './types';
 
 export type CastAttributes = {
     [key: string]: StringConstructor | NumberConstructor | BooleanConstructor | ((value: any) => any) | (new(value: any) => any);
@@ -13,8 +9,8 @@ function isConstructor(obj: any): obj is (new(value: any) => any) {
     return !! obj.prototype && !! obj.prototype.constructor.name;
 }
 
-export class Model implements JsonApiResource {
-    public type: string;
+export class Model<Type extends string = any> implements JsonApiResource<Type> {
+    public type: Type;
     public id: string;
     public attributes: KeyValueObject = {};
     public relationships: JsonApiRelationships = {};
@@ -25,7 +21,7 @@ export class Model implements JsonApiResource {
 
     [field: string]: any;
 
-    constructor(data: JsonApiResource, protected store: Store) {
+    constructor(data: JsonApiResource<Type>, protected store: Store) {
         this.merge(data);
     }
 
@@ -45,13 +41,15 @@ export class Model implements JsonApiResource {
     }
 
     public getRelationship(name: string): any {
-        return this.store.find(this.relationships[name].data as JsonApiIdentifier);
+        const data = this.relationships[name].data;
+        // https://github.com/microsoft/TypeScript/issues/14107
+        return Array.isArray(data) ? this.store.find(data) : this.store.find(data);
     }
 
     /**
      * Make a resource identifier object for this model.
      */
-    identifier(): JsonApiIdentifier {
+    identifier(): JsonApiIdentifier<Type> {
         return {
             id: this.id,
             type: this.type
@@ -61,9 +59,9 @@ export class Model implements JsonApiResource {
     /**
      * Merge new JSON:API resource data into the model.
      */
-    merge(data: JsonApiResource): void {
+    merge(data: Partial<JsonApiResource<Type>>): void {
         if ('type' in data) {
-            this.type = data.type;
+            this.type = data.type as Type;
         }
 
         if ('id' in data) {
@@ -74,28 +72,40 @@ export class Model implements JsonApiResource {
             Object.assign(this.attributes, data.attributes);
 
             Object.keys(data.attributes).forEach(name => {
-                if (! Object.getOwnPropertyDescriptor(Object.getPrototypeOf(this), name)) {
-                    Object.defineProperty(this, name, {
-                        configurable: true,
-                        get: () => this.getAttribute(name),
-                    });
-                }
+                if (
+                    Object.getOwnPropertyDescriptor(Object.getPrototypeOf(this), name)
+                    || Object.getOwnPropertyDescriptor(this, name)
+                ) return;
+
+                this[name] = null;
+
+                Object.defineProperty(this, name, {
+                    get: () => this.getAttribute(name),
+                    configurable: true,
+                    enumerable: true,
+                });
             });
         }
 
         if ('relationships' in data) {
-            for (const [name, relationship] of Object.entries(data.relationships)) {
+            Object.entries(data.relationships).forEach(([name, relationship]) => {
                 this.relationships[name] = this.relationships[name] || {};
 
                 Object.assign(this.relationships[name], relationship);
 
-                if (! Object.getOwnPropertyDescriptor(Object.getPrototypeOf(this), name)) {
-                    Object.defineProperty(this, name, {
-                        configurable: true,
-                        get: () => this.getRelationship(name),
-                    });
-                }
-            }
+                if (
+                    Object.getOwnPropertyDescriptor(Object.getPrototypeOf(this), name)
+                    || Object.getOwnPropertyDescriptor(this, name)
+                ) return;
+
+                this[name] = null;
+
+                Object.defineProperty(this, name, {
+                    get: () => this.getRelationship(name),
+                    configurable: true,
+                    enumerable: true,
+                });
+            });
         }
 
         if ('links' in data) {