+++ src/types.ts
@@ -2,19 +2,19 @@ export type KeyValueObject = {
     [key: string]: any;
 }
 
-export interface JsonApiDocument {
-    data: JsonApiResource | JsonApiResource[];
+export interface JsonApiDocument<Type extends string = string> {
+    data: JsonApiResource<Type> | JsonApiResource<Type>[];
     included?: JsonApiResource[];
     meta?: KeyValueObject;
     links?: KeyValueObject;
 }
 
-export interface JsonApiIdentifier {
-    type: string;
+export interface JsonApiIdentifier<Type extends string = string> {
+    type: Type;
     id: string;
 }
 
-export interface JsonApiResource extends JsonApiIdentifier {
+export interface JsonApiResource<Type extends string = string> extends JsonApiIdentifier<Type> {
     attributes?: KeyValueObject;
     relationships?: JsonApiRelationships;
     meta?: KeyValueObject;
@@ -25,8 +25,16 @@ export interface JsonApiRelationships {
     [relationName: string]: JsonApiRelationship;
 }
 
-export interface JsonApiRelationship {
-    data?: JsonApiIdentifier | JsonApiIdentifier[];
+export interface JsonApiRelationship<Type extends string = string> {
+    data?: JsonApiIdentifier<Type> | JsonApiIdentifier<Type>[] | null;
     meta?: KeyValueObject;
     links?: KeyValueObject;
+}
+
+export interface JsonApiRelationshipToOne<Type extends string = string> extends JsonApiRelationship<Type> {
+    data?: JsonApiIdentifier<Type> | null;
+}
+
+export interface JsonApiRelationshipToMany<Type extends string = string> extends JsonApiRelationship<Type> {
+    data?: JsonApiIdentifier<Type>[];
 }