+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support model type inference:
+```ts
+const store = new Store({ users: User });
+store.find('users', '1') // User
+```
+
+### Removed
+- Remove the `Store.model` method. Models must now be defined when the Store is constructed.
 
 ## [0.1.0-beta.7] - 2021-09-01
 ### Fixed